backend/llamaPython.py
from typing import Any
from llama_cpp import Llama

text="Q: Name the planets in the solar system? A: "

class Llama2 :
    def __init__(self) :
        self.llm = Llama(model_path="./models/llama-2-7b-chat.ggmlv3.q4_0.bin",)

    def __call__(self,prompt):
        output = self.llm( prompt=prompt,max_tokens=1000, echo=True)
        return output

    